MAINT: Split QueryParser
split concerns of query parser into QuerySorter and QueryGrouper each responsible for specific parsing methods
